Move OnboardingStepper to its sole consumer, embedding-hub
