+++ frontend/src/metabase/embedding/embedding-hub/components/OnboardingStepper/OnboardingStepper.module.css
