+++ frontend/src/metabase/embedding/embedding-hub/components/OnboardingStepper/OnboardingStepper.tsx
