+++ frontend/src/metabase/embedding/embedding-hub/components/OnboardingStepper/OnboardingStepper.unit.spec.tsx
