+++ frontend/src/metabase/embedding/embedding-hub/components/OnboardingStepper/OnboardingStepperContext.tsx
