+++ frontend/src/metabase/embedding/embedding-hub/components/OnboardingStepper/OnboardingStepperStep.tsx
