+++ frontend/src/metabase/embedding/embedding-hub/components/OnboardingStepper/hooks/use-auto-advance-step.ts
