+++ frontend/src/metabase/embedding/embedding-hub/components/OnboardingStepper/hooks/use-scroll-step-into-view.ts
