+++ frontend/src/metabase/embedding/embedding-hub/components/OnboardingStepper/index.ts
