+++ frontend/src/metabase/embedding/embedding-hub/components/OnboardingStepper/types.ts
