+++ frontend/src/metabase/embedding/embedding-hub/components/OnboardingStepper/utils/get-next-step-to-activate.ts
