+++ frontend/src/metabase/embedding/embedding-hub/components/SetupPermissionsAndTenantsPage/SetupPermissionsAndTenantsPage.tsx
@@ -3,8 +3,8 @@ import { match } from "ts-pattern";
 import { t } from "ttag";
 
 import { useGetEmbeddingHubChecklistQuery } from "metabase/api/embedding-hub";
-import { OnboardingStepper } from "metabase/common/components/OnboardingStepper";
-import type { OnboardingStepperHandle } from "metabase/common/components/OnboardingStepper/types";
+import { OnboardingStepper } from "metabase/embedding/embedding-hub/components/OnboardingStepper";
+import type { OnboardingStepperHandle } from "metabase/embedding/embedding-hub/components/OnboardingStepper/types";
 import {
   type CreatedTenantData,
   PLUGIN_TENANTS,
